refactor(backend): split god files by domain entity
- Split operations.go into user_store.go, plant_store.go, sensor_store.go, task_store.go
- Split models.go into user.go, plant.go, task.go
- Extract jsonError helper into handlers/helpers.go
- Rename PlantService → SensorService (accurately reflects sensor ingestion business logic)
- Update base.go and ingest.go references accordingly
- No business logic changes — pure file organization
