+++ backend/server/internal/handlers/upload.go
@@ -1,0 +1,61 @@
+package handlers
+
+import (
+	"encoding/json"
+	"fmt"
+	"io"
+	"net/http"
+	"os"
+)
+
+// HandleUploadImage handles multipart image file uploads.
+func (h *Handler) HandleUploadImage(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodPost {
+		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
+		return
+	}
+
+	if err := r.ParseMultipartForm(10 << 20); err != nil {
+		http.Error(w, "File too large", http.StatusBadRequest)
+		return
+	}
+
+	file, handler, err := r.FormFile("file")
+	if err != nil {
+		http.Error(w, "Invalid file", http.StatusBadRequest)
+		return
+	}
+	defer func() {
+		if err := file.Close(); err != nil {
+			fmt.Printf("error closing file: %v\n", err)
+		}
+	}()
+
+	// Create uploads folder if not exists
+	if err := os.MkdirAll("./uploads", os.ModePerm); err != nil {
+		http.Error(w, "Failed to create folder", http.StatusInternalServerError)
+		return
+	}
+
+	dst, err := os.Create("./uploads/" + handler.Filename)
+	if err != nil {
+		http.Error(w, "Failed to save file", http.StatusInternalServerError)
+		return
+	}
+	defer func() {
+		if err := dst.Close(); err != nil {
+			fmt.Printf("error closing dst file: %v\n", err)
+		}
+	}()
+
+	if _, err := io.Copy(dst, file); err != nil {
+		http.Error(w, "Failed to write file", http.StatusInternalServerError)
+		return
+	}
+
+	url := "/uploads/" + handler.Filename
+	w.Header().Set("Content-Type", "application/json")
+	if err := json.NewEncoder(w).Encode(map[string]string{"url": url}); err != nil {
+		fmt.Printf("error encoding json response: %v\n", err)
+	}
+}
