+++ backend/server/internal/storage/task_store.go
@@ -5,167 +5,14 @@ import (
 	"errors"
 
 	"plantbee-backend/internal/models"
+
 	"github.com/lib/pq"
 )
 
+// ErrTaskAlreadyActive is returned when trying to create a duplicate active task.
 var ErrTaskAlreadyActive = errors.New("a task of this type is already active for this plant")
 
-// User operations
-func (d *DB) UpsertUser(user *models.User) error {
-	query := `
-		INSERT INTO users (intra_id, email, login, image_url, intend_to_help, first_visit, water_count, logged_in)
-		VALUES ($1, $2, $3, $4, $5, $6, $7, true)
-		ON CONFLICT (intra_id) DO UPDATE
-		SET email = EXCLUDED.email,
-			login = EXCLUDED.login,
-			image_url = EXCLUDED.image_url,
-			logged_in = true
-		RETURNING id, intend_to_help, first_visit, water_count;
-	`
-	return d.QueryRow(
-		query,
-		user.IntraID,
-		user.Email,
-		user.Login,
-		user.ImageURL,
-		user.IntendToHelp,
-		user.FirstVisit,
-		user.WaterCount,
-	).Scan(&user.ID, &user.IntendToHelp, &user.FirstVisit, &user.WaterCount)
-}
-
-func (d *DB) GetUserByID(userID int) (*models.User, error) {
-	query := `SELECT id, email, login, image_url, intend_to_help, first_visit, water_count FROM users WHERE id = $1`
-	var user models.User
-	err := d.QueryRow(query, userID).Scan(
-		&user.ID,
-		&user.Email,
-		&user.Login,
-		&user.ImageURL,
-		&user.IntendToHelp,
-		&user.FirstVisit,
-		&user.WaterCount,
-	)
-	if err != nil {
-		return nil, err
-	}
-	return &user, nil
-}
-
-func (d *DB) CompleteWelcome(userID int, intendToHelp bool) error {
-	query := `UPDATE users SET intend_to_help = $1, first_visit = false WHERE id = $2`
-	_, err := d.Exec(query, intendToHelp, userID)
-	return err
-}
-
-func (d *DB) SetUserIntention(userID int, help bool) error {
-	query := `UPDATE users SET intend_to_help = $1 WHERE id = $2`
-	_, err := d.Exec(query, help, userID)
-	return err
-}
-
-func (d *DB) IncrementWaterCount(userID int) error {
-	query := `UPDATE users SET water_count = water_count + 1 WHERE id = $1`
-	_, err := d.Exec(query, userID)
-	return err
-}
-
-func (d *DB) SetUserLoggedOut(userID int) error {
-	query := `UPDATE users SET logged_in = false WHERE id = $1`
-	_, err := d.Exec(query, userID)
-	return err
-}
-
-// Sensor Operations
-func (d *DB) SaveSensorReadings(reading *models.SensorReading) error {
-	query := `
-		INSERT INTO sensor_readings (sensor_id, moisture, wake_time, battery_level)
-		VALUES ($1, $2, $3, $4)
-		RETURNING id, recorded_at;
-	`
-	return d.QueryRow(
-		query,
-		reading.SensorID,
-		reading.Moisture,
-		reading.WakeTime,
-		reading.BatteryLevel,
-	).Scan(&reading.ID, &reading.RecordedAt)
-}
-
-// Plant Operations
-func (d *DB) CreatePlant(plant *models.Plant) error {
-	query := `
-		INSERT INTO plants (name, species, category, pot_volume_liters, light_need, target_moisture, sensor_id, image_url, owner_id)
-		VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)
-		RETURNING id, created_at;
-	`
-	return d.QueryRow(query,
-		plant.Name, plant.Species, plant.Category,
-		plant.PotVolumeLiters, plant.LightRequirement,
-		plant.TargetMoisture, plant.SensorID, plant.ImageURL, plant.OwnerID,
-	).Scan(&plant.ID, &plant.CreatedAt)
-}
-
-func (d *DB) GetPlantBySensorID(sensorID string) (*models.Plant, error) {
-	query := `SELECT id, name, target_moisture, pot_volume_liters FROM plants WHERE sensor_id = $1`
-
-	var plant models.Plant
-	err := d.QueryRow(query, sensorID).Scan(
-		&plant.ID,
-		&plant.Name,
-		&plant.TargetMoisture,
-		&plant.PotVolumeLiters,
-	)
-
-	return &plant, err
-}
-
-func (d *DB) GetPlantByOwnerID(ownerID int) (*models.Plant, error) {
-	query := `SELECT id, name, target_moisture FROM plants WHERE owner_id = $1`
-	var plant models.Plant
-	err := d.QueryRow(query, ownerID).Scan(&plant.ID, &plant.Name, &plant.TargetMoisture)
-	return &plant, err
-}
-
-// GetAllPlants returns all plants with owner names for the plant list page.
-func (d *DB) GetAllPlants() ([]models.PlantListItem, error) {
-	query := `
-		SELECT p.id, p.name, p.light_need, p.target_moisture, p.current_moisture, p.image_url, COALESCE(u.login, '') AS owner_name
-		FROM plants p
-		LEFT JOIN users u ON p.owner_id = u.id
-		ORDER BY p.created_at DESC
-	`
-
-	rows, err := d.Query(query)
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		_ = rows.Close()
-	}()
-
-	var plants []models.PlantListItem
-	for rows.Next() {
-		var p models.PlantListItem
-		if err := rows.Scan(&p.ID, &p.Name, &p.LightRequirement, &p.TargetMoisture, &p.CurrentMoisture, &p.ImageURL, &p.OwnerName); err != nil {
-			return nil, err
-		}
-		plants = append(plants, p)
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-
-	// Return empty slice instead of nil so JSON encodes as [] not null
-	if plants == nil {
-		plants = []models.PlantListItem{}
-	}
-
-	return plants, nil
-}
-
-// Task Operations
+// CreateTask inserts a new task into the database.
 func (d *DB) CreateTask(task *models.Task) error {
 	query := `
         INSERT INTO tasks (plant_id, type, water_amount, message, status, scheduled_at)
